+++ src/coverage/__tests__/format.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 
-const { format } = require('../format-coverage')
+const { format } = require('../format')
 
 describe('format', () => {
   it('formats report', () => {
@@ -15,10 +15,32 @@ describe('format', () => {
     )).toMatchSnapshot()
   })
 
+  it('formats report with multiple folders links', () => {
+    expect(format(
+      require('./basic-report4.json'),
+      undefined,
+      'http://example.com/artifacts'
+    )).toMatchSnapshot()
+  })
+
   it('formats delta report', () => {
     expect(format(
       require('./basic-report2.json'),
       require('./basic-report.json')
+    )).toMatchSnapshot()
+  })
+
+  it('formats delta report 2', () => {
+    expect(format(
+      require('./basic-report3.json'),
+      require('./basic-report2.json')
+    )).toMatchSnapshot()
+  })
+
+  it('formats delta report with no changes', () => {
+    expect(format(
+      require('./basic-report2.json'),
+      require('./basic-report2.json')
     )).toMatchSnapshot()
   })