+++ changelog.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - UNRELEASED
+
+### Changes
+- New comment format
+- Use markdown in comment where possible
+- Added link to project in comments
+- New Emojis to cover more scenarios
+
+### Fixes
+- No longer return 1 when file is empty/skipped
+- Fix file artifact links
+- Fix non-workflow build artifact discovery
+- Fix computation of artifact links with multiple folders
+
+### Misc
+- More code coverage/tests
+
+
 ## [1.0.0] - 2017-12-19
 
 - Initial release